requirements.txt
numpy==1.24.2
openai==0.27.2
python-dotenv==1.0.0
tiktoken==0.3.1
tqdm==4.65.0
